Migration of methods from ParserStorageUtil
