+++ src/main/java/duke/logic/parsers/Parser.java
@@ -2,8 +2,6 @@ package duke.logic.parsers;
 
 import duke.commons.Messages;
 import duke.commons.exceptions.ParseException;
-import duke.commons.exceptions.StartEndDateBeforeNowException;
-import duke.commons.exceptions.StartEndDateDiscordException;
 import duke.logic.commands.AddSampleItineraryCommand;
 import duke.logic.commands.Command;
 import duke.logic.commands.EditorCommand;
@@ -18,26 +16,26 @@ import duke.logic.commands.RouteNodeShowCommand;
 import duke.logic.commands.RouteShowCommand;
 import duke.logic.commands.ShowItineraryCommand;
 import duke.logic.commands.ViewScheduleCommand;
-import duke.logic.parsers.commandparser.AddEventParser;
-import duke.logic.parsers.commandparser.AddProfileParser;
-import duke.logic.parsers.commandparser.CreateNewItineraryParser;
-import duke.logic.parsers.commandparser.DeleteParser;
-import duke.logic.parsers.commandparser.DoneParser;
-import duke.logic.parsers.commandparser.FindParser;
-import duke.logic.parsers.commandparser.GetBusRouteParser;
-import duke.logic.parsers.commandparser.GetBusStopParser;
-import duke.logic.parsers.commandparser.LocationSearchParser;
-import duke.logic.parsers.commandparser.QuickEditParser;
-import duke.logic.parsers.commandparser.RecommendationParser;
-import duke.logic.parsers.commandparser.RouteAddParser;
-import duke.logic.parsers.commandparser.RouteDeleteParser;
-import duke.logic.parsers.commandparser.RouteEditParser;
-import duke.logic.parsers.commandparser.RouteListParser;
-import duke.logic.parsers.commandparser.RouteNodeAddParser;
-import duke.logic.parsers.commandparser.RouteNodeDeleteParser;
-import duke.logic.parsers.commandparser.RouteNodeEditParser;
-import duke.logic.parsers.commandparser.RouteNodeListParser;
-import duke.logic.parsers.commandparser.StaticMapParser;
+import duke.logic.parsers.commandParsers.AddEventParser;
+import duke.logic.parsers.commandParsers.AddProfileParser;
+import duke.logic.parsers.commandParsers.CreateNewItineraryParser;
+import duke.logic.parsers.commandParsers.DeleteParser;
+import duke.logic.parsers.commandParsers.DoneParser;
+import duke.logic.parsers.commandParsers.FindParser;
+import duke.logic.parsers.commandParsers.GetBusRouteParser;
+import duke.logic.parsers.commandParsers.GetBusStopParser;
+import duke.logic.parsers.commandParsers.LocationSearchParser;
+import duke.logic.parsers.commandParsers.QuickEditParser;
+import duke.logic.parsers.commandParsers.RecommendationParser;
+import duke.logic.parsers.commandParsers.RouteAddParser;
+import duke.logic.parsers.commandParsers.RouteDeleteParser;
+import duke.logic.parsers.commandParsers.RouteEditParser;
+import duke.logic.parsers.commandParsers.RouteListParser;
+import duke.logic.parsers.commandParsers.RouteNodeAddParser;
+import duke.logic.parsers.commandParsers.RouteNodeDeleteParser;
+import duke.logic.parsers.commandParsers.RouteNodeEditParser;
+import duke.logic.parsers.commandParsers.RouteNodeListParser;
+import duke.logic.parsers.commandParsers.StaticMapParser;
 
 /**
  * Parser for commands entered by the user. It reads from standard input and returns Command objects.
