+++ src/main/java/duke/logic/parsers/commandParsers/AddEventParser.java
@@ -1,4 +1,4 @@
-package duke.logic.parsers.commandparser;
+package duke.logic.parsers.commandParsers;
 
 import duke.commons.Messages;
 import duke.commons.exceptions.ApiException;
